Fix encoding length types
The encoding traits are attempting to convert `std::ptrdiff_t`, which
are signed values, as `std::size_t` in the `ucapacity` type, which is
unsigned. This was leading to warnings about signed to unsigned
conversions.

This change explicitly casts the distance to `std::size_t` to
silence those warnings -- since the distances cannot be negative in
this context.

Change-Category: bugfix

+++ lib/alloy-core/include/alloy/core/string/encoding/encoding_traits.hpp
@@ -596,7 +596,7 @@ auto alloy::core::encoding_traits<Encoding>::length(
     return count;
   } else {
     // Non-multi-unit codepoints can be computed by counting distance
-    return uquantity<char32>{std::distance(first, last)};
+    return uquantity<char32>{static_cast<std::size_t>(std::distance(first, last))};
   }
 }
